Re organize js to accomodate new ExtractionApp
