+++ app/frontend/js/apps/TransformationApp/JumpTo.jsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { selectRawRecord } from "/js/features/RawRecordSlice";
+import { selectRawRecord } from "/js/features/TransformationApp/RawRecordSlice";
 import {
   clickedOnRunFields,
   selectAppDetails,
-} from "/js/features/AppDetailsSlice";
-import { selectFieldIds } from "/js/features/FieldsSlice";
+} from "/js/features/TransformationApp/AppDetailsSlice";
+import { selectFieldIds } from "/js/features/TransformationApp/FieldsSlice";
 
 const JumpTo = () => {
   const dispatch = useDispatch();
