+++ app/frontend/js/apps/TransformationApp/components/FieldNavigationPanel.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { selectFieldIds, selectAllFields } from "~/js/features/FieldsSlice";
+import { selectFieldIds, selectAllFields } from "~/js/features/TransformationApp/FieldsSlice";
 import { filter, concat } from "lodash";
 import FieldNavigationListItem from "./FieldNavigationListItem";
 import AddField from "~/js/apps/TransformationApp/components/AddField";
-import { toggleDisplayFields } from "~/js/features/UiFieldsSlice";
+import { toggleDisplayFields } from "~/js/features/TransformationApp/UiFieldsSlice";
 
 const FieldNavigationPanel = () => {
   const dispatch = useDispatch();
