+++ app/frontend/js/features/TransformationApp/AppDetailsSlice.js
