+++ app/frontend/js/features/TransformationApp/UiAppDetailsSlice.js
